1. until will fix KinClient then currently when restore success then create a new KinClient in order for him to be up to date and pass it to the caller. 2. add progress bar for balance and public address fields 3. clear all account when starting the sample app.

+++ kin-recovery/recovery-sample/src/main/java/recovery/sample/AccountCreator.java
@@ -17,7 +17,7 @@ import okhttp3.Response;
 
 public class AccountCreator {
 
-	private static final int FUND_KIN_AMOUNT = 100;
+	private static final double FUND_KIN_AMOUNT = 100;
 	private static final String URL_CREATE_ACCOUNT =
 		"https://friendbot-testnet.kininfrastructure.com?addr=%s&amount=" + String.valueOf(FUND_KIN_AMOUNT);
 	private final OkHttpClient okHttpClient;
